Fix more decimal/double type conversions
- Fixed PatternEngine: 7 IFeatureBus.Publish calls now cast double to decimal
- Fixed EnhancedBayesianPriors: Updated namespace reference from BotCore.Math to BotCore.DecimalMath.MathHelper

Remaining: 66 CS errors in FeaturePublisher, ZoneFeatureResolvers, UnifiedBarPipeline, EpochFreezeEnforcement, SafeHoldDecisionPolicy

+++ src/BotCore/Patterns/PatternEngine.cs
@@ -73,9 +73,9 @@ public class PatternEngine
                     });
 
                     // Publish individual pattern metrics to feature bus
-                    _featureBus.Publish(symbol, now, $"pattern.kind::{detector.PatternName}", result.Score);
-                    _featureBus.Publish(symbol, now, $"pattern.direction::{detector.PatternName}", result.Direction);
-                    _featureBus.Publish(symbol, now, $"pattern.confidence::{detector.PatternName}", result.Confidence);
+                    _featureBus.Publish(symbol, now, $"pattern.kind::{detector.PatternName}", (decimal)result.Score);
+                    _featureBus.Publish(symbol, now, $"pattern.direction::{detector.PatternName}", (decimal)result.Direction);
+                    _featureBus.Publish(symbol, now, $"pattern.confidence::{detector.PatternName}", (decimal)result.Confidence);
                 }
             }
             catch (InvalidOperationException ex)
@@ -92,8 +92,8 @@ public class PatternEngine
         var patternScores = AggregateScores(results);
 
         // Publish summary scores to feature bus  
-        _featureBus.Publish(symbol, now, "pattern.bull_score", patternScores.BullScore);
-        _featureBus.Publish(symbol, now, "pattern.bear_score", patternScores.BearScore);
+        _featureBus.Publish(symbol, now, "pattern.bull_score", (decimal)patternScores.BullScore);
+        _featureBus.Publish(symbol, now, "pattern.bear_score", (decimal)patternScores.BearScore);
         _featureBus.Publish(symbol, now, "pattern.total_count", patternScores.PatternFlags.Count);
 
         _logger.LogDebug("Pattern analysis for {Symbol}: Bull={BullScore:F3}, Bear={BearScore:F3}, Patterns={Count}",
@@ -138,9 +138,9 @@ public class PatternEngine
             
             // Publish pattern scores to feature bus
             var timestamp = DateTime.UtcNow;
-            _featureBus.Publish(symbol, timestamp, "pattern.bull_score", detailsResult.BullScore);
-            _featureBus.Publish(symbol, timestamp, "pattern.bear_score", detailsResult.BearScore);
-            _featureBus.Publish(symbol, timestamp, "pattern.confidence", detailsResult.OverallConfidence);
+            _featureBus.Publish(symbol, timestamp, "pattern.bull_score", (decimal)detailsResult.BullScore);
+            _featureBus.Publish(symbol, timestamp, "pattern.bear_score", (decimal)detailsResult.BearScore);
+            _featureBus.Publish(symbol, timestamp, "pattern.confidence", (decimal)detailsResult.OverallConfidence);
 
             return detailsResult;
         }
